+++ src/main/kotlin/com/github/reonaore/fuzzyfinderintellijplugin/actions/OpenLiveGrepAction.kt
@@ -1,0 +1,23 @@
+package com.github.reonaore.fuzzyfinderintellijplugin.actions
+
+import com.github.reonaore.fuzzyfinderintellijplugin.MyBundle
+import com.github.reonaore.fuzzyfinderintellijplugin.ui.LiveGrepDialog
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAwareAction
+
+class OpenLiveGrepAction : DumbAwareAction() {
+
+    override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
+
+    override fun update(e: AnActionEvent) {
+        e.presentation.text = MyBundle.message("action.openLiveGrep.text")
+        e.presentation.description = MyBundle.message("action.openLiveGrep.description")
+        e.presentation.isEnabled = e.project != null
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        LiveGrepDialog(project).show()
+    }
+}