+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # FuzzyFinderIntellijPlugin Changelog
 
 ## [Unreleased]
+### Added
+- Live Grep dialog powered by `rg` with smart-case regex matching
+- Configurable executable path for `rg`
 
 ## [0.1.0] - 2026-04-23
 ### Added